requirements.txt
anyio==3.7.0
arviz==0.12.1
cachetools==5.3.1
certifi==2023.5.7
cftime==1.6.2
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
deprecat==2.1.1
dill==0.3.6
exceptiongroup==1.1.1
fastapi==0.97.0
fastprogress==1.0.3
filelock==3.12.2
fonttools==4.40.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
netCDF4==1.6.4
numpy==1.22.1
packaging==23.1
pandas==2.0.2
patsy==0.5.3
Pillow==9.5.0
pydantic==1.10.9
pymc3==3.11.5
pyparsing==3.1.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
scipy==1.7.3
seaborn==0.12.2
semver==3.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
Theano-PyMC==1.1.2
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
wrapt==1.15.0
xarray==2023.5.0
xarray-einstats==0.5.1
zipp==3.15.0
